# pkg/api/proto/pvz/v1/pvz/v1/pvz.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pvz/v1/pvz.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PVZ",
      "description": "PVZ service example",
      "externalDocs": {
        "description": "Find out more about grpc-gateway",
        "url": "https://github.com/grpc-ecosystem/grpc-gateway?tab=readme-ov-file"
      }
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/add-order": {
      "post": {
        "operationId": "PVZ_AddOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Описание для AddOrderRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pvzAddOrderRequest"
            }
          }
        ],
        "tags": [
          "PVZ"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/delete-order": {
      "delete": {
        "operationId": "PVZ_DeleteOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pvzDeleteOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ID.id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "PVZ"
        ]
      }
    },
    "/v1/list-order": {
      "get": {
        "operationId": "PVZ_ListOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pvzListOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "PVZ"
        ]
      }
    },
    "/v1/list-order/{id.id}": {
      "get": {
        "operationId": "PVZ_FindOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pvzFindOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id.id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "PVZ"
        ]
      }
    },
    "/v1/list-refund": {
      "get": {
        "operationId": "PVZ_ListRefund",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pvzListRefundResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "pagination.pageSize",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "PVZ"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "pvzAddOrderRequest": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/pvzOrder"
        }
      },
      "description": "Описание для AddOrderRequest",
      "title": "AddOrderRequest информация",
      "required": [
        "order"
      ]
    },
    "pvzDeleteOrderResponse": {
      "type": "object"
    },
    "pvzFindOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/pvzOrderInfo"
        }
      }
    },
    "pvzID": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pvzListOrderResponse": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pvzOrderInfo"
          }
        }
      }
    },
    "pvzListRefundResponse": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pvzOrderInfo"
          }
        }
      }
    },
    "pvzOrder": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "idReceiver": {
          "type": "string",
          "format": "int64"
        },
        "storageTime": {
          "type": "string",
          "format": "date-time"
        },
        "weight": {
          "type": "number",
          "format": "double"
        },
        "packaging": {
          "type": "string"
        }
      }
    },
    "pvzOrderInfo": {
      "type": "object",
      "properties": {
        "order": {
          "type": "string"
        }
      }
    },
    "pvzPagination": {
      "type": "object",
      "properties": {
        "page": {
          "type": "string",
          "format": "int64"
        },
        "pageSize": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}